feat!: Working MacOS (arm64 only) and Linux (amd64 + arm64) CI (#1)
* ci for other platforms

* ci for other platforms

* ci for other platforms

* remove mac amd64

* tailwind

* reward hacking

* mac build fixes

* env issue

* .

* more polyfill

* more ci pain

* fixes for 9.2 tag

* more tweaks

* tweak

* ci hell

* new session!

* static

* lots of fixes

---------

+++ crates/foyer-backend-stub/src/regions.rs
@@ -86,7 +86,7 @@ fn synthesize_for(track_id: &EntityId) -> Vec<Region> {
             id: EntityId::new(format!("region.{slug}.{i}")),
             track_id: track_id.clone(),
             name: format!("{slug} {}", i + 1),
-            start_samples: start,
+            start_samples: start as i64,
             length_samples: dur,
             color: None,
             muted: false,
